+++ tests/main.rs
@@ -262,13 +262,37 @@ macro_rules! test_transpose {
     }
 }
 
+macro_rules! test_mul_add {
+    ($([$ty:ident;$len:literal] as $target:ident),+) => {
+        $(
+            let alpha : [$ty;$len] = random();
+            let beta : [$ty;$len] = random();
+            let gamma : [$ty;$len] = random();
+
+            let naive = (0usize..$len).into_iter()
+                .map(|i| alpha[i].mul_add(beta[i], gamma[i]));
+
+            let alpha = <$target>::new(alpha);
+            let beta = <$target>::new(beta);
+            let gamma = <$target>::new(gamma);
+            let simd = alpha.mul_add(beta, gamma);
+
+            Into::<[$ty;$len]>::into(simd)
+                .into_iter()
+                .zip(naive)
+                //.for_each(|(i, (simd, naive))| panic!("{} * {} + {} = {:?} or {:?}", alpha[i], beta[i], gamma[i], simd, naive));
+                .for_each(|(simd, naive)| assert!((simd - naive).abs() <= $ty::EPSILON * 2., concat!("mul_add for ", stringify!($target))));
+        )*
+    }
+}
+
 test_other!(
     add, sub, mul, div, 
     min as vmin, max as vmax
 );
 
 test_horiz!(
-    min, max, add as sum
+    min, max, add as sum, mul as prod
 );
 
 test_mappings!(
@@ -461,4 +485,39 @@ pub fn zip () {
     );
 }
 
-// vtrn1q_f32
+#[test]
+pub fn fma () {
+    test_mul_add!(
+        [f32;2] as f32x2,
+        [f32;4] as f32x4,
+        [f32;6] as f32x6,
+        [f32;8] as f32x8,
+        [f32;10] as f32x10,
+        [f32;12] as f32x12,
+        [f32;14] as f32x14,
+        [f32;16] as f32x16,
+    
+        [f64;2] as f64x2,
+        [f64;4] as f64x4,
+        [f64;6] as f64x6,
+        [f64;8] as f64x8,
+        [f64;10] as f64x10,
+        [f64;12] as f64x12,
+        [f64;14] as f64x14,
+        [f64;16] as f64x16
+    );
+}
+
+#[test]
+fn mul_add () {
+    let array = [core::f32::consts::FRAC_1_PI, 0.0012];
+    let simd = f32x2::new(array);
+    let simd = simd.mul_add(simd, simd);
+
+    let reg = array.into_iter()
+        .map(|x| x.mul_add(x, x))
+        .enumerate()
+        .for_each(|(i, x)| {
+            assert_eq!(x, simd[i]);
+        });
+}